Refactor(#56): Schedule 패키지 분리
기존에 Calendar에 있던 내용을 Schedule패키지를 새로 만들어서 분리
