+++ application/wypl-core/src/main/java/com/wypl/wyplcore/calendar/service/CalendarService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.wypl.googleoauthclient.domain.AuthMember;
 import com.wypl.jpacalendardomain.calendar.domain.Calendar;
 import com.wypl.jpacalendardomain.calendar.domain.MemberCalendar;
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
-import com.wypl.jpacalendardomain.calendar.repository.ScheduleRepository;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.reopository.ScheduleRepository;
 import com.wypl.jpamemberdomain.member.domain.Member;
 import com.wypl.wyplcore.calendar.data.DateSearchCondition;
 import com.wypl.wyplcore.calendar.data.request.CalendarFindRequest;
