+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/data/response/ScheduleFindResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 
 public record ScheduleFindResponse(
