+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/repetition/RepetitionStrategyFactory.java
@@ -3,7 +3,7 @@ package com.wypl.wyplcore.schedule.service.repetition;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.wypl.jpacalendardomain.calendar.data.RepetitionCycle;
+import com.wypl.jpacalendardomain.schedule.data.RepetitionCycle;
 import com.wypl.wyplcore.schedule.service.repetition.strategy.DayRepetitionStrategy;
 import com.wypl.wyplcore.schedule.service.repetition.strategy.MonthRepetitionStrategy;
 import com.wypl.wyplcore.schedule.service.repetition.strategy.RepetitionStrategy;
