+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/repetition/strategy/DayRepetitionStrategy.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
 
 public class DayRepetitionStrategy implements RepetitionStrategy {
