+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/repetition/strategy/MonthRepetitionStrategy.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 import java.time.Period;
 import java.util.List;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
 
 public class MonthRepetitionStrategy implements RepetitionStrategy {
