+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/repetition/strategy/RepetitionStrategy.java
@@ -3,7 +3,7 @@ package com.wypl.wyplcore.schedule.service.repetition.strategy;
 import java.time.LocalDate;
 import java.util.List;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
 
 public interface RepetitionStrategy {
