+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/repetition/strategy/WeekRepetitionStrategy.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
 
 public class WeekRepetitionStrategy implements RepetitionStrategy {
