+++ application/wypl-core/src/test/java/com/wypl/wyplcore/schedule/service/repetition/DayRepetitionStrategyTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.ScheduleFixture;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
