+++ application/wypl-core/src/test/java/com/wypl/wyplcore/schedule/service/repetition/YearRepetitionStrategyTest.java
@@ -13,7 +13,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 import com.wypl.wyplcore.schedule.data.response.ScheduleFindResponse;
 import com.wypl.wyplcore.schedule.service.repetition.strategy.YearRepetitionStrategy;
