+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/domain/Schedule.java
@@ -1,4 +1,4 @@
-package com.wypl.jpacalendardomain.calendar.domain;
+package com.wypl.jpacalendardomain.schedule.domain;
 
 import java.time.Duration;
 import java.time.LocalDate;
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 import org.hibernate.annotations.SQLRestriction;
 
-import com.wypl.jpacalendardomain.calendar.data.RepetitionCycle;
+import com.wypl.jpacalendardomain.schedule.data.RepetitionCycle;
 import com.wypl.jpacommon.JpaBaseEntity;
 
 import jakarta.persistence.Column;
