+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/mapper/ScheduleMapper.java
@@ -1,8 +1,8 @@
-package com.wypl.jpacalendardomain.calendar.mapper;
+package com.wypl.jpacalendardomain.schedule.mapper;
 
-import com.wypl.jpacalendardomain.calendar.data.ConvertibleSchedule;
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
-import com.wypl.jpacalendardomain.calendar.domain.ScheduleInfo;
+import com.wypl.jpacalendardomain.schedule.data.ConvertibleSchedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.ScheduleInfo;
 
 public class ScheduleMapper {
