+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/reopository/ScheduleInfoRepository.java
@@ -1,8 +1,8 @@
-package com.wypl.jpacalendardomain.calendar.repository;
+package com.wypl.jpacalendardomain.schedule.reopository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.wypl.jpacalendardomain.calendar.domain.ScheduleInfo;
+import com.wypl.jpacalendardomain.schedule.domain.ScheduleInfo;
 
 public interface ScheduleInfoRepository extends JpaRepository<ScheduleInfo, Long> {
 }
