+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/reopository/ScheduleRepository.java
@@ -1,8 +1,8 @@
-package com.wypl.jpacalendardomain.calendar.repository;
+package com.wypl.jpacalendardomain.schedule.reopository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.wypl.jpacalendardomain.calendar.domain.Schedule;
+import com.wypl.jpacalendardomain.schedule.domain.Schedule;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long>, ScheduleRepositoryCustom {
 }
